Support module hot-reload (#20)
Code swap technique implies unloading the old module code by setting
`package.loaded[module] = nil` and subsequently `require()` it again.

To make it smooth and to avoid status intervention one has to persist
some data across reload though - the socket object, members, events,
running fibers. But fiber code should be updated too. So there is a new
Lua module stash which is used to store all those things. A fiber body
with an infinite loop is implemented with the only closure on that
stash, and the real logic is stored there.